requirements.txt
tensorflow==2.20.0
streamlit==1.50.0
numpy==2.0.2
pillow==10.4.0
